Next Major Iteration
After several more hours of collaborating, this is the new iteration. It includes the first prototype of the turn code that has begun testing and more refinements on the course correction code. Next is to rework the course correction system to be separated into individual states as it has gotten quite bulky.