Completed First Forward() Prototype
Continued work with Dave on resolving the issues related to driving in a straight line. We measured the motor speed difference between the two sides to add some software calibration (noting that at the upper end of the range the relationship becomes nonlinear). He also reworked the course correction to be simplified and more reliable based on the issues the prior iteration faced. The result is the robot can now drive forward in a mostly straight line and if angled too far towards a wall, it can also correct away from an approaching wall without driving into any walls. I worked on documentation and rewrote the turn handling code to do a turn on the spot rather than an arcing turn to help increase reliability in the turn and ease difficulties in calibration. The new turn system has not yet been tested.